requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.1
billiard==3.6.1.0
certifi==2019.9.11
chardet==3.0.4
cssutils==1.0.2
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-ckeditor==5.8.0
django-contrib-comments==1.9.1
django-debug-toolbar==2.1
django-js-asset==1.2.2
django-sendgrid-v5==0.8.0
Django-tinymce-filebrowser==0.2.1
future==0.18.2
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
lxml==4.4.1
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.2.1
pynliner==0.8.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.0
six==1.13.0
sorl-thumbnail==12.5.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
zipp==0.6.0
